separating browser side helpers from general helpers, adding jest and some test cases, adding test command to the pre-commit hooks
