+++ .husky/pre-commit
@@ -3,3 +3,4 @@
 
 yarn lint
 yarn lint-staged
+yarn test
