+++ jest.config.js
@@ -1,0 +1,9 @@
+const { compilerOptions } = require('./tsconfig.json');
+
+/** @type {import('ts-jest').JestConfigWithTsJest} */
+module.exports = {
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+  roots: ['<rootDir>'],
+  modulePaths: [compilerOptions.baseUrl],
+};
