+++ package.json
@@ -6,7 +6,9 @@
   "author": "",
   "license": "MIT",
   "devDependencies": {
+    "@jest/globals": "^29.7.0",
     "@types/chrome": "^0.0.237",
+    "@types/jest": "^29.5.11",
     "@types/lodash": "^4.14.202",
     "@types/web": "^0.0.130",
     "@typescript-eslint/eslint-plugin": "^6.13.2",
@@ -15,9 +17,11 @@
     "copy-webpack-plugin": "^11.0.0",
     "html-webpack-plugin": "^5.5.2",
     "husky": "^8.0.3",
+    "jest": "^29.7.0",
     "lint-staged": "^13.2.2",
     "prettier": "2.8.8",
     "run-script-os": "^1.1.6",
+    "ts-jest": "^29.1.1",
     "ts-loader": "^9.4.3",
     "tsconfig-paths-webpack-plugin": "^4.1.0",
     "typescript": "^5.1.3",
@@ -50,6 +54,7 @@
     "package": "web-ext build",
     "prepare": "husky install",
     "lint": "yarn eslint src/**/*.ts",
+    "test": "jest",
     "win-build-gecko": "Remove-Item dist -Recurse -Force; webpack --mode development --config engines/gecko/webpack.config.js; node scripts\\post-build-update-manifest.js",
     "win-build-chromium": "Remove-Item dist -Recurse -Force; webpack --mode development --config engines/chromium/webpack.config.js; node scripts\\post-build-update-manifest.js",
     "nix-build-gecko": "rm -rf dist && webpack --mode development --config engines/gecko/webpack.config.js && node scripts/post-build-update-manifest.js",
